pyproject.toml
[project]
name = "quant-auto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioredis>=2.0.1",
    "alpaca-py>=0.43.2",
    "asyncpg>=0.31.0",
    "langchain-anthropic>=1.3.3",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-redis>=0.3.5",
    "loguru>=0.7.3",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pytz>=2025.2",
    "signalrcore>=1.0.1",
    "sqlmodel>=0.0.37",
    "websockets>=16.0",
]
